order_services/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
twilio